+++ core/src/main/kotlin/sift/core/api/ScopeTrail.kt
@@ -11,6 +11,10 @@ internal class ScopeTrail : Sequence<Element> {
 
     private val elements: MutableList<Element> = mutableListOf()
 
+    fun intersects(other: ScopeTrail): Boolean {
+        return elements.intersect(other.elements.toSet()).isNotEmpty()
+    }
+
     operator fun plus(element: Element): ScopeTrail {
         return ScopeTrail(elements + element)
     }
@@ -21,6 +25,10 @@ internal class ScopeTrail : Sequence<Element> {
 
     operator fun contains(other: ScopeTrail): Boolean {
         return other.elements.containsAll(elements)
+    }
+
+    operator fun contains(element: Element): Boolean {
+        return element in elements
     }
 
     override fun toString(): String {