+++ core/src/test/kotlin/sift/core/api/testdata/KnownBugs.kt
@@ -5,9 +5,6 @@ import org.junit.jupiter.api.Disabled
 import org.junit.jupiter.api.Test
 import org.objectweb.asm.tree.ClassNode
 import sift.core.api.*
-import sift.core.api.Dsl.classes
-import sift.core.asm.classNode
-import sift.core.entity.Entity
 import sift.core.entity.EntityService
 
 @Disabled
@@ -17,31 +14,6 @@ class KnownBugs {
         debugLog = true
     }
 
-    @Test
-    fun `explode Payload in List field and associate property from the main class`(){
-
-        val payload = Entity.Type("payload")
-
-        classes {
-            fields {
-                signature {
-                    typeArguments {
-                        explodeType(synthesize = true) {
-                            entity(payload)
-                        }
-                    }
-                }
-            }
-            property(payload, "field-owner", readName())
-        }.execute(listOf(classNode(FieldClass::class))) { es ->
-            val entities = es[payload].values
-            assertThat(entities).hasSize(1)
-
-            entities.first().let { e ->
-                assertThat(e["field-owner"]).isEqualTo(listOf("FieldClass"))
-            }
-        }
-    }
 
     private fun Action<Unit, Unit>.execute(
             cns: List<ClassNode>,
@@ -54,8 +26,3 @@ class KnownBugs {
     }
 }
 
-private class FieldClass {
-    val payloads: List<PayLoad> = listOf()
-}
-
-private class PayLoad